pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llama-cpp-python-gradio"
version = "0.0.4"
description = "A Python package for creating Gradio applications with llama.cpp models"
authors = [
    { name = "AK", email = "ahsen.khaliq@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "gradio>=5.5.0",
    "llama-cpp-python",
]

[project.urls]
homepage = "https://github.com/AK391/llama-cpp-python-gradio"
repository = "https://github.com/AK391/llama-cpp-python-gradio"

[project.optional-dependencies]
dev = ["pytest"]

[tool.hatch.build.targets.wheel]
packages = ["llama_cpp_python_gradio"]

